+++ src/App.tsx
@@ -1,12 +1,10 @@
 import React from 'react';
 import './App.scss';
-import ClassNames from 'classnames';
-
 import todosFromServer from './api/todos';
 import usersFromServer from './api/users';
-
 import { User } from './types/User';
 import { Todo } from './types/Todo';
+import { TodoList } from './components/TodoList';
 
 function getUser(userId: number): User | null {
   const foundUser = usersFromServer.find(user => user.id === userId);
@@ -19,44 +17,6 @@ export const todos: Todo[] = todosFromServer.map(todo => ({
   ...todo,
   user: getUser(todo.userId),
 }));
-
-type UserInfoProps = {
-  user: User | null;
-};
-
-const UserInfo: React.FC<UserInfoProps> = ({ user }) => (
-  <a className="UserInfo" href={`mailto:${user?.email}`}>
-    {user?.name}
-  </a>
-);
-
-type TodoInfoProps = {
-  todo: Todo;
-};
-
-const TodoInfo: React.FC<TodoInfoProps> = ({ todo }) => (
-  <article className={ClassNames('TodoInfo',
-    {
-      'TodoInfo--completed': !todo.completed,
-    })}
-  >
-    <h2 className="TodoInfo__title">{todo.title}</h2>
-
-    <UserInfo user={todo.user} />
-  </article>
-);
-
-type Props = {
-  todolist: Todo[];
-};
-
-const TodoList: React.FC<Props> = ({ todolist }) => {
-  return (
-    todolist.map(todo => (
-      <TodoInfo todo={todo} />
-    ))
-  );
-};
 
 export const App: React.FC = () => (
   <div className="App">
