+++ src/components/TodoList/TodoList.tsx
@@ -1,4 +1,14 @@
-// Add the required types and props
-export const TodoList = () => (
-  <>TodoList markup</>
-);
+import { Todo } from '../../types/Todo';
+import { TodoInfo } from '../TodoInfo';
+
+type Props = {
+  todolist: Todo[];
+};
+
+export const TodoList: React.FC<Props> = ({ todolist }) => {
+  return (
+    todolist.map(todo => (
+      <TodoInfo todo={todo} key={todo.id} />
+    ))
+  );
+};
